voicebot/emotion_recognition_service/requirements.txt
numpy ==1.18.5
absl-py==0.13.0
appdirs==1.4.4
astor==0.8.1
audioread==2.1.9
beautifulsoup4==4.9.3
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
decorator==5.0.9
Flask==1.1.2
gast==0.5.0
google-pasta==0.2.0
grpcio==1.38.1
h5py
idna==2.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
librosa==0.8.1
llvmlite==0.36.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
numba==0.53.1
packaging==21.0
pooch==1.4.0
protobuf==3.17.3
pycparser==2.20
pyparsing==2.4.7
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
aeneas==1.7.3.0